refactor: AIDDサイクル定義を整理
工程責務とGoal遷移をworkflowへ集約し、重複checkerをrendererへ統合する。
