+++ .agents/skills/aidd-cycle/references/artifact-validation.md
@@ -1,0 +1,117 @@
+# AIDD Artifact Validation
+
+Use this reference only for Requirements and Design phase execution.
+
+## Structural Model
+
+Managed AIDD artifacts use schema version 2.
+
+- Goal inputs: temporary `requirements_goal` or `design_goal` JSON.
+- Canonical artifacts:
+  `docs/ai-driven-development/workspaces/<workspace>/requirements.json` and
+  `design-doc.json`.
+- Display outputs: `requirements.md` and `design-doc.md`, generated by
+  `render_aidd_artifact.py`.
+
+The validators enforce canonical paths, regular non-symlink inputs, Issue
+snapshot identity, full inventories, baseline hashes, rule dependencies,
+transition status, block references, evidence roles, and evidence owners.
+Schema v2 additionally enforces canonical headings, substantive non-placeholder
+text, unique evidence references, and unambiguous evidence-to-owner mapping.
+
+## Workspace
+
+```sh
+python3 .agents/skills/aidd-cycle/scripts/validate_workspace.py \
+  --repo-root <repo-root> \
+  --issue <owner/repo#number> \
+  --workspace <workspace>
+```
+
+## Requirements Goal
+
+Save the exact fetched Issue body and proposed Goal JSON in temporary regular
+files. Run both commands before `create_goal`:
+
+```sh
+python3 .agents/skills/aidd-cycle/scripts/validate_requirements_goal.py \
+  --issue <owner/repo#number> --issue-url <canonical-url> \
+  --issue-updated-at <updatedAt> --issue-body <issue-body-file> \
+  --document <requirements-goal-json> --rule-map docs/harness/rule-map.json \
+  --repo-root <repo-root> --kind goal
+
+python3 .agents/skills/aidd-cycle/scripts/validate_requirements_continuity.py \
+  --issue <owner/repo#number> --issue-body <issue-body-file> \
+  --document <requirements-goal-json> --kind goal \
+  --repo-root <repo-root> --workspace <workspace>
+```
+
+The Goal must snapshot the Issue, select at least one direct rule, include its
+declared dependency closure, inventory every current and baseline requirement
+and required section, and classify explicit retirements. Evidence quotes must
+come from the snapshotted Issue, occur in their owned content, and be unique per
+changed or new requirement and per changed or new section. Retirement evidence
+must name the ID and explicitly state retirement without negating it.
+
+## Requirements Artifact
+
+Render the canonical Markdown, then rerun both gates with the retained Goal:
+
+```sh
+python3 .agents/skills/aidd-cycle/scripts/render_aidd_artifact.py \
+  --source <requirements-json> --output <requirements-md> \
+  --repo-root <repo-root>
+
+python3 .agents/skills/aidd-cycle/scripts/validate_requirements_goal.py \
+  --issue <owner/repo#number> --issue-url <canonical-url> \
+  --issue-updated-at <updatedAt> --issue-body <issue-body-file> \
+  --document <requirements-json> --rule-map docs/harness/rule-map.json \
+  --repo-root <repo-root> --kind artifact \
+  --goal-document <requirements-goal-json>
+
+python3 .agents/skills/aidd-cycle/scripts/validate_requirements_continuity.py \
+  --issue <owner/repo#number> --issue-body <issue-body-file> \
+  --document <requirements-json> --kind artifact \
+  --repo-root <repo-root> --workspace <workspace> \
+  --goal-document <requirements-goal-json>
+```
+
+Fetch the Issue again before completing the phase. Its identity, `updatedAt`,
+body, and hash must still match the Goal snapshot.
+
+## Design Goal and Artifact
+
+Revalidate the canonical Requirements artifact first. Then use the Design gate
+for the temporary Goal JSON and again for the canonical artifact:
+
+```sh
+python3 .agents/skills/aidd-cycle/scripts/validate_design_coverage.py \
+  --issue <owner/repo#number> --issue-url <canonical-url> \
+  --issue-updated-at <updatedAt> --issue-body <issue-body-file> \
+  --rule-map docs/harness/rule-map.json \
+  --requirements <canonical-requirements-json> \
+  --document <design-goal-json> --kind goal \
+  --repo-root <repo-root> --workspace <workspace>
+
+python3 .agents/skills/aidd-cycle/scripts/render_aidd_artifact.py \
+  --source <design-json> --output <design-md> --repo-root <repo-root>
+
+python3 .agents/skills/aidd-cycle/scripts/validate_design_coverage.py \
+  --issue <owner/repo#number> --issue-url <canonical-url> \
+  --issue-updated-at <updatedAt> --issue-body <issue-body-file> \
+  --rule-map docs/harness/rule-map.json \
+  --requirements <canonical-requirements-json> \
+  --document <design-json> --kind artifact \
+  --repo-root <repo-root> --workspace <workspace>
+```
+
+Each requirement must own one design evidence reference and one verification
+evidence reference. Each replaced baseline section must own baseline evidence.
+Hashes decide preserved versus replaced; headings remain display labels.
+
+## Repository-Wide Display Check
+
+```sh
+python3 .agents/skills/aidd-cycle/scripts/render_aidd_artifact.py \
+  --repo-root <repo-root> --check-all
+```
